moved XdsLoadReportCallback interface into XdsLoadReportClient
