+++ xds/src/main/java/io/grpc/xds/XdsLoadReportClientImpl.java
@@ -346,23 +346,6 @@ final class XdsLoadReportClientImpl implements XdsLoadReportClient {
     }
   }
 
-  /**
-   * Callbacks for passing information received from client load reporting responses to xDS load
-   * balancer, such as the load reporting interval requested by the traffic director.
-   *
-   * <p>Implementations are not required to be thread-safe as callbacks will be invoked in xDS load
-   * balancer's {@link SynchronizationContext}.
-   */
-  interface XdsLoadReportCallback {
-
-    /**
-     * The load reporting interval has been received.
-     *
-     * @param reportIntervalNano load reporting interval requested by remote traffic director.
-     */
-    void onReportResponse(long reportIntervalNano);
-  }
-
   abstract static class XdsLoadReportClientFactory {
 
     private static final XdsLoadReportClientFactory DEFAULT_INSTANCE =
